requirements.txt
requests
torch
sounddevice
coqui-tts
faster-whisper
pyaudio
numpy
soundfile